Empirical Validation of Performance Metrics for Web Page
Under the average website load scenario, numerous users use their
browsers to surf certain websites. A single web page presented in a
browser can generate tens, sometimes hundreds of unique HTTP requests.
The browser, after receiving all of the responses renders the page for
the user to view. Every browser (IE, Firefox, Chrome, etc.) has its own
way of generating the HTTP request according to the viewed web page. In
a load scenario, the overall number of HTTP requests is directly related
to the number of users that are surfing concurrently. All requests hit
the server during the test generating a response for each request.

The goal of performance testing is to evaluate the application’s
performance with respect to real world scenarios. The issues which must
be addressed are that the performance of the system during peak hours
i.e, response time, reliability, availability etc., points at which the
system performance degrades or system fails as well as the impact of the
degraded performance on the customer loyalty, sales and profits.
Following metrics parameters are used for the study:
Average Response Time
Peak Response Time
No. of Hits
Latency
Throughput